chore (refs T22614): move news lib to client/js.
