+++ client/js/bundles/news/newsAdminEdit.js
@@ -16,7 +16,7 @@ import DpEditor from '@DpJs/components/core/DpEditor/DpEditor'
 import DpUploadFiles from '@DpJs/components/core/DpUpload/DpUploadFiles'
 import dpValidate from '@DpJs/lib/validation/dpValidate'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
-import newsAdminInit from '@DemosPlanNewsBundle/lib/newsAdmin'
+import newsAdminInit from '@DpJs/lib/news/newsAdmin'
 
 const components = {
   DpChangeStateAtDate,
