+++ client/js/lib/news/newsAdmin.js
